+++ packages/runtime/runtime-definitions/src/dataStoreFactory.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { IFluidDataStoreContext } from "./componentContext";
+import { IFluidDataStoreContext } from "./dataStoreContext";
 
 declare module "@fluidframework/core-interfaces" {
     // eslint-disable-next-line @typescript-eslint/no-empty-interface
@@ -17,18 +17,18 @@ export interface IProvideFluidDataStoreFactory {
 }
 
 /**
- * IFluidDataStoreFactory create components.  It is associated with an identifier (its `type` member)
- * and usually provided to consumers using this mapping through a component registry.
+ * IFluidDataStoreFactory create data stores.  It is associated with an identifier (its `type` member)
+ * and usually provided to consumers using this mapping through a data store registry.
  */
 export interface IFluidDataStoreFactory extends IProvideFluidDataStoreFactory {
     /**
-     * String that uniquely identifies the type of component created by this factory.
+     * String that uniquely identifies the type of data store created by this factory.
      */
     type?: string;
 
     /**
-     * Generates runtime for the component from the component context. Once created should be bound to the context.
-     * @param context - Context for the component.
+     * Generates runtime for the data store from the data store context. Once created should be bound to the context.
+     * @param context - Context for the data store.
      */
     instantiateDataStore(context: IFluidDataStoreContext): void;
 }
