+++ packages/runtime/runtime-definitions/src/dataStoreRegistry.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { IProvideFluidDataStoreFactory } from "./componentFactory";
+import { IProvideFluidDataStoreFactory } from "./dataStoreFactory";
 
 declare module "@fluidframework/core-interfaces" {
     // eslint-disable-next-line @typescript-eslint/no-empty-interface
@@ -11,7 +11,7 @@ declare module "@fluidframework/core-interfaces" {
 }
 
 /**
- * A single registry entry that may be used to create components
+ * A single registry entry that may be used to create data stores
  */
 export type FluidDataStoreRegistryEntry =
     Readonly<Partial<IProvideFluidDataStoreRegistry & IProvideFluidDataStoreFactory>>;
@@ -32,8 +32,8 @@ export interface IProvideFluidDataStoreRegistry {
 }
 
 /**
- * An association of identifiers to component registry entries, where the
- * entries can be used to create components.
+ * An association of identifiers to data store registry entries, where the
+ * entries can be used to create data stores.
  */
 export interface IFluidDataStoreRegistry extends IProvideFluidDataStoreRegistry {
     get(name: string): Promise<FluidDataStoreRegistryEntry | undefined>;
